# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mystic Weave",
    "description": "Persistent game state backend for a text-based RPG powered by a custom GPT. The GPT is the narrator; this API is the memory. d100 roll-under resolution with domain scores and competency tiers.",
    "version": "5.3.1"
  },
  "servers": [
    {
      "url": "https://mysticweave-production.up.railway.app",
      "description": "Production (Railway)"
    }
  ],
  "paths": {
    "/state/{session_id}": {
      "get": {
        "summary": "Load State",
        "description": "Load the full game state for a session. Returns 404 if not found.",
        "operationId": "load_state_state__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Save State",
        "description": "Save game state, deep-merge character and world updates onto stored state, and append one log entry atomically.",
        "operationId": "save_state_state__session_id__post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveStateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/state/{session_id}/delta": {
      "post": {
        "summary": "Save State Delta",
        "description": "Apply a validated structured state delta to stored game state, then append one log entry atomically.",
        "operationId": "save_state_delta_state__session_id__delta_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyStateDeltaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/state/{session_id}/annotation": {
      "post": {
        "summary": "Annotate State",
        "description": "Append a canon-correction annotation to the session log without mutating gameplay state. Recorded as an admin_correction TypedLogEntry prefixed with its category.",
        "operationId": "annotate_state_state__session_id__annotation_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnnotationRequest": {
        "properties": {
          "annotation": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Annotation"
          },
          "category": {
            "type": "string",
            "enum": [
              "canon_correction",
              "operational_constraint",
              "rule_clarification",
              "narrator_correction"
            ],
            "title": "Category"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "annotation",
          "category"
        ],
        "title": "AnnotationRequest",
        "description": "Body for POST /state/{session_id}/annotation.\n\nA canon correction or operational note that does NOT mutate gameplay state.\nStored as a TypedLogEntry of type 'admin_correction' on the session log.\nUse this for factual/canon corrections, session-local operational notes,\nrule clarifications, and narrator self-corrections — anything where the\ngameplay /delta endpoint would reject the payload as a no-op."
      },
      "AnnotationResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "annotation_id": {
            "type": "string",
            "title": "Annotation Id"
          },
          "appended_to_log": {
            "type": "boolean",
            "title": "Appended To Log",
            "default": true
          },
          "log_entry_index": {
            "type": "integer",
            "title": "Log Entry Index"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "annotation_id",
          "log_entry_index"
        ],
        "title": "AnnotationResponse",
        "description": "Response for POST /state/{session_id}/annotation."
      },
...
      "ApplyStateDeltaRequest": {
        "properties": {
          "character": {
            "$ref": "#/components/schemas/CharacterStateDelta"
          },
          "world": {
            "$ref": "#/components/schemas/WorldStateDelta"
          },
          "log_entry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/TypedLogEntry"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Entry"
          },
          "time_elapsed": {
            "$ref": "#/components/schemas/TimeElapsed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApplyStateDeltaRequest",
        "description": "Body for POST /state/{session_id}/delta"
      },
...
      "GameStateResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "character": {
            "$ref": "#/components/schemas/CharacterModel-Output"
          },
          "world": {
            "$ref": "#/components/schemas/WorldModel-Output"
          },
          "log": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Log"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "character",
          "world",
          "log"
        ],
        "title": "GameStateResponse",
        "description": "Response for GET /state/{session_id}"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SaveStateRequest": {
        "properties": {
          "character": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CharacterModel-Input"
              },
              {
                "$ref": "#/components/schemas/CharacterStateDelta"
              }
            ],
            "title": "Character"
          },
          "world": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorldModel-Input"
              },
              {
                "$ref": "#/components/schemas/WorldStateDelta"
              }
            ],
            "title": "World"
          },
          "log_entry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/TypedLogEntry"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Entry"
          },
          "time_elapsed": {
            "$ref": "#/components/schemas/TimeElapsed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SaveStateRequest",
        "description": "Body for POST /state/{session_id}"
      },
...
    }
  }
}